requirements.txt
pyenchant
Flask